feat: Add exam dependency to DI container
This commit adds the necessary code changes to include the `exam.dependency.dart` file in the DI container. The `setupExamDependency` function is called in the `setupLocator` function to register the `ExamApiService` and `ExamRepo` as lazy singletons. This change allows for the proper dependency injection of exam-related classes throughout the application.